chore: some updates for documentation [B1M2-3511]
